+++ doc/rst/MochiKit/VersionHistory.rst
@@ -25,6 +25,8 @@
 2009-XX-YY      v1.4.3 (bug fix release)
 
 - Fixed MochiKit.Logging usage of map without a namespace (#338).
+- Fixed MochiKit.Color.Color.prototype.isLight() and isDark() to not
+  return constant false and true values (#341).
 
 2008-11-27      v1.4.2 (bug fix release)